* Fixed broken tests
