+++ content/forgero-vanilla/src/test/java/com/sigmundgranaas/forgero/test/attribute/tool/AxeTest.java
@@ -1,6 +1,6 @@
-package com.sigmundgranaas.forgeroforge.test.attribute.tool;
+package com.sigmundgranaas.forgero.test.attribute.tool;
 
-import static com.sigmundgranaas.forgeroforge.test.util.StateHelper.state;
+import static com.sigmundgranaas.forgero.test.util.StateHelper.state;
 
 import com.sigmundgranaas.forgero.core.property.v2.attribute.attributes.AttackDamage;
 import com.sigmundgranaas.forgero.core.property.v2.attribute.attributes.AttackSpeed;
@@ -8,83 +8,83 @@ import com.sigmundgranaas.forgero.core.property.v2.attribute.attributes.Durabili
 import com.sigmundgranaas.forgero.core.property.v2.attribute.attributes.MiningSpeed;
 import com.sigmundgranaas.forgero.core.property.v2.attribute.attributes.Rarity;
 import com.sigmundgranaas.forgero.core.state.composite.Construct;
-import com.sigmundgranaas.forgeroforge.test.util.AttributeTester;
-import com.sigmundgranaas.forgeroforge.test.util.ForgeroPackageTest;
+import com.sigmundgranaas.forgero.test.util.AttributeTester;
+import com.sigmundgranaas.forgero.test.util.ForgeroPackageTest;
 import org.junit.jupiter.api.Test;
 
 /**
  * All vanilla tools should stay within some distance to their minecraft counterparts
  */
-public class PickaxeTest extends ForgeroPackageTest {
+public class AxeTest extends ForgeroPackageTest {
 	@Test
-	void testOakPickaxe() {
-		AttributeTester.tester(state("forgero:oak-pickaxe"))
+	void testOakAxe() {
+		AttributeTester.tester(state("forgero:oak-axe"))
 				.add(Durability.KEY, 50, 10)
 				.add(MiningSpeed.KEY, 2, 1)
 				.add(Rarity.KEY, 20, 5)
-				.add(AttackDamage.KEY, 2, 0.5)
-				.add(AttackSpeed.KEY, -2.0, 0.5)
+				.add(AttackDamage.KEY, 6, 0.5)
+				.add(AttackSpeed.KEY, -3.0, 0.5)
 				.run();
 	}
 
 	@Test
-	void testIronPickaxe() {
-		AttributeTester.tester(state("forgero:iron-pickaxe"))
+	void testIronAxe() {
+		AttributeTester.tester(state("forgero:iron-axe"))
 				.add(Durability.KEY, 250, 50)
 				.add(MiningSpeed.KEY, 6, 1)
 				.add(Rarity.KEY, 40, 10)
-				.add(AttackDamage.KEY, 4, 0.5)
-				.add(AttackSpeed.KEY, -2.0, 0.5)
+				.add(AttackDamage.KEY, 8, 0.5)
+				.add(AttackSpeed.KEY, -3.0, 0.5)
 				.run();
 	}
 
 	@Test
-	void testGoldPickaxe() {
-		AttributeTester.tester(state("forgero:gold-pickaxe"))
+	void testGoldAxe() {
+		AttributeTester.tester(state("forgero:gold-axe"))
 				.add(Durability.KEY, 41, 10)
 				.add(MiningSpeed.KEY, 12, 1)
 				.add(Rarity.KEY, 60, 10)
-				.add(AttackDamage.KEY, 2, 0.5)
-				.add(AttackSpeed.KEY, -2.0, 0.5)
+				.add(AttackDamage.KEY, 6, 0.5)
+				.add(AttackSpeed.KEY, -3.0, 0.5)
 				.run();
 	}
 
 	@Test
-	void testDiamondPickaxe() {
-		AttributeTester.tester(state("forgero:diamond-pickaxe"))
+	void testDiamondAxe() {
+		AttributeTester.tester(state("forgero:diamond-axe"))
 				.add(Durability.KEY, 1500, 300)
 				.add(MiningSpeed.KEY, 8, 1)
 				.add(Rarity.KEY, 80, 10)
-				.add(AttackDamage.KEY, 5, 0.5)
-				.add(AttackSpeed.KEY, -2.0, 0.5)
+				.add(AttackDamage.KEY, 9, 0.5)
+				.add(AttackSpeed.KEY, -3.0, 0.5)
 				.run();
 	}
 
 	@Test
-	void testNetheritePickaxe() {
-		AttributeTester.tester(state("forgero:netherite-pickaxe"))
+	void testNetheriteAxe() {
+		AttributeTester.tester(state("forgero:netherite-axe"))
 				.add(Durability.KEY, 2100, 300)
 				.add(MiningSpeed.KEY, 9, 1)
 				.add(Rarity.KEY, 110, 10)
-				.add(AttackDamage.KEY, 6, 1.5)
-				.add(AttackSpeed.KEY, -2.0, 0.5)
+				.add(AttackDamage.KEY, 10, 1.5)
+				.add(AttackSpeed.KEY, -3.0, 0.5)
 				.run();
 	}
 
 	@Test
-	void testDiamondPickaxeIronHandle() {
+	void testDiamondAxeIronHandle() {
 		var pickaxe = Construct.builder()
-				.id("forgero:diamond-pickaxe")
+				.id("forgero:diamond-axe")
 				.addIngredient(state("forgero:iron-handle"))
-				.addIngredient(state("forgero:diamond-pickaxe_head"))
+				.addIngredient(state("forgero:diamond-axe_head"))
 				.build();
 
 		AttributeTester.tester(pickaxe)
 				.add(Durability.KEY, 1600, 200)
 				.add(MiningSpeed.KEY, 9, 1)
 				.add(Rarity.KEY, 100, 20)
-				.add(AttackDamage.KEY, 6, 1.5)
-				.add(AttackSpeed.KEY, -2.0, 0.4)
+				.add(AttackDamage.KEY, 8, 1.5)
+				.add(AttackSpeed.KEY, -2.8, 0.4)
 				.run();
 	}
 }
