+++ content/forgero-vanilla/src/test/java/com/sigmundgranaas/forgero/test/util/AttributeTester.java
@@ -1,4 +1,4 @@
-package com.sigmundgranaas.forgeroforge.test.util;
+package com.sigmundgranaas.forgero.test.util;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
