+++ content/forgero-vanilla/src/test/java/com/sigmundgranaas/forgero/test/util/ForgeroPackageTest.java
@@ -1,10 +1,11 @@
-package com.sigmundgranaas.forgeroforge.test.util;
+package com.sigmundgranaas.forgero.test.util;
 
 import org.junit.jupiter.api.BeforeAll;
 
 public class ForgeroPackageTest {
 	@BeforeAll
 	public static void runSetup() {
+		StateHelper.ruleSetup();
 		ForgeroPipeLineSetup.setup();
 	}
 }
