+++ content/forgero-vanilla/src/test/java/com/sigmundgranaas/forgero/test/util/StateHelper.java
@@ -1,21 +1,19 @@
-package com.sigmundgranaas.forgero.testutil;
+package com.sigmundgranaas.forgero.test.util;
 
-import java.util.Set;
-import java.util.function.Supplier;
-
+import com.sigmundgranaas.forgero.core.ForgeroStateRegistry;
 import com.sigmundgranaas.forgero.core.api.identity.DefaultRules;
 import com.sigmundgranaas.forgero.core.api.identity.ModificationRuleRegistry;
 import com.sigmundgranaas.forgero.core.api.identity.sorting.SortingRule;
 import com.sigmundgranaas.forgero.core.api.identity.sorting.SortingRuleRegistry;
-import com.sigmundgranaas.forgero.core.property.attribute.Category;
-import com.sigmundgranaas.forgero.core.state.Ingredient;
-import com.sigmundgranaas.forgero.core.state.NameCompositor;
-import com.sigmundgranaas.forgero.core.state.composite.Construct;
-import com.sigmundgranaas.forgero.core.state.upgrade.slot.EmptySlot;
+import com.sigmundgranaas.forgero.core.state.State;
 import com.sigmundgranaas.forgero.core.type.Type;
 
-public class Tools {
-	public static NameCompositor defaultCompositor() {
+public class StateHelper {
+	public static State state(String id) {
+		return ForgeroStateRegistry.stateFinder().find(id).orElseThrow();
+	}
+
+	public static void ruleSetup() {
 		SortingRuleRegistry sorting = SortingRuleRegistry.staticRegistry();
 		sorting.registerRule("forgero:schematic", SortingRule.of(Type.SCHEMATIC, 20));
 		sorting.registerRule("forgero:material", SortingRule.of(Type.MATERIAL, 10));
@@ -30,14 +28,5 @@ public class Tools {
 		modification.registerRule("forgero:hoe", DefaultRules.hoe.build());
 		modification.registerRule("forgero:axe", DefaultRules.axe.build());
 		modification.registerRule("forgero:shovel", DefaultRules.shovel.build());
-		return new NameCompositor(modification, sorting);
 	}
-
-	public static Supplier<Construct> IRON_PICKAXE = () -> Construct.builder()
-			.addIngredient(Ingredient.of(ToolParts.PICKAXE_HEAD))
-			.addIngredient(Ingredient.of(ToolParts.HANDLE))
-			.addUpgrade(new EmptySlot(0, Type.BINDING, "", Set.of(Category.UTILITY)))
-			.type(Type.PICKAXE)
-			.compositor(defaultCompositor())
-			.build();
 }
