--- content/forgero-vanilla/src/test/java/com/sigmundgranaas/forgeroforge/test/util/StateHelper.java
@@ -1,11 +1,0 @@
-package com.sigmundgranaas.forgeroforge.test.util;
-
-import com.sigmundgranaas.forgero.core.ForgeroStateRegistry;
-import com.sigmundgranaas.forgero.core.state.State;
-
-public class StateHelper {
-	public static State state(String id) {
-		return ForgeroStateRegistry.stateFinder().find(id).orElseThrow();
-	}
-
-}
