+++ forgero-core/src/main/java/com/sigmundgranaas/forgero/core/state/composite/BaseComposite.java
@@ -218,6 +218,11 @@ public abstract class BaseComposite implements Composite {
 			return (B) this;
 		}
 
+		public B compositor(NameCompositor compositor) {
+			this.compositor = compositor;
+			return (B) this;
+		}
+
 		protected void compositeName() {
 			if (this.name == null && !ingredientList.isEmpty()) {
 				this.name = compositor.compositeName(ingredientList);
