+++ forgero-core/src/test/java/com/sigmundgranaas/forgero/HandleStateTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.Test;
 
 public class HandleStateTest {
 
+
 	@Test
 	void testCreateHandleState() {
 		var handle = HANDLE;
