+++ forgero-core/src/test/java/com/sigmundgranaas/forgero/testutil/Name.java
@@ -1,20 +1,13 @@
 package com.sigmundgranaas.forgero.testutil;
 
-import java.util.Set;
-import java.util.function.Supplier;
-
 import com.sigmundgranaas.forgero.core.api.identity.DefaultRules;
 import com.sigmundgranaas.forgero.core.api.identity.ModificationRuleRegistry;
 import com.sigmundgranaas.forgero.core.api.identity.sorting.SortingRule;
 import com.sigmundgranaas.forgero.core.api.identity.sorting.SortingRuleRegistry;
-import com.sigmundgranaas.forgero.core.property.attribute.Category;
-import com.sigmundgranaas.forgero.core.state.Ingredient;
 import com.sigmundgranaas.forgero.core.state.NameCompositor;
-import com.sigmundgranaas.forgero.core.state.composite.Construct;
-import com.sigmundgranaas.forgero.core.state.upgrade.slot.EmptySlot;
 import com.sigmundgranaas.forgero.core.type.Type;
 
-public class Tools {
+public class Name {
 	public static NameCompositor defaultCompositor() {
 		SortingRuleRegistry sorting = SortingRuleRegistry.staticRegistry();
 		sorting.registerRule("forgero:schematic", SortingRule.of(Type.SCHEMATIC, 20));
@@ -32,12 +25,4 @@ public class Tools {
 		modification.registerRule("forgero:shovel", DefaultRules.shovel.build());
 		return new NameCompositor(modification, sorting);
 	}
-
-	public static Supplier<Construct> IRON_PICKAXE = () -> Construct.builder()
-			.addIngredient(Ingredient.of(ToolParts.PICKAXE_HEAD))
-			.addIngredient(Ingredient.of(ToolParts.HANDLE))
-			.addUpgrade(new EmptySlot(0, Type.BINDING, "", Set.of(Category.UTILITY)))
-			.type(Type.PICKAXE)
-			.compositor(defaultCompositor())
-			.build();
 }
