+++ forgero-core/src/test/java/com/sigmundgranaas/forgero/testutil/ToolParts.java
@@ -2,6 +2,7 @@ package com.sigmundgranaas.forgero.testutil;
 
 import static com.sigmundgranaas.forgero.testutil.Materials.IRON;
 import static com.sigmundgranaas.forgero.testutil.Materials.OAK;
+import static com.sigmundgranaas.forgero.testutil.Name.defaultCompositor;
 
 import java.util.Collections;
 import java.util.List;
@@ -16,6 +17,8 @@ public class ToolParts {
 			.addIngredient(Schematics.HANDLE_SCHEMATIC)
 			.addUpgrades(EmptySlot.of(List.of(Types.MATERIAL, Types.GEM), Collections.emptySet()))
 			.type(Type.HANDLE)
+			.compositor(defaultCompositor())
+
 			.build();
 
 	public static Construct REFINED_HANDLE = Construct.builder()
@@ -23,18 +26,21 @@ public class ToolParts {
 			.addIngredient(Schematics.REFINED_HANDLE_SCHEMATIC)
 			.addUpgrades(EmptySlot.of(List.of(Types.MATERIAL, Types.GEM, Types.MATERIAL), Collections.emptySet()))
 			.type(Type.HANDLE)
+			.compositor(defaultCompositor())
 			.build();
 
 	public static Construct PICKAXE_HEAD = Construct.builder()
 			.addIngredient(IRON)
 			.addIngredient(Schematics.PICKAXE_HEAD_SCHEMATIC)
 			.type(Type.PICKAXE_HEAD)
+			.compositor(defaultCompositor())
 			.build();
 
 	public static Construct SWORD_BLADE = Construct.builder()
 			.addIngredient(IRON)
 			.addIngredient(Schematics.SWORD_BLADE_SCHEMATIC)
 			.type(Type.SWORD_BLADE)
+			.compositor(defaultCompositor())
 			.build();
 
 
@@ -42,17 +48,20 @@ public class ToolParts {
 			.addIngredient(IRON)
 			.addIngredient(Schematics.MASTERCRAFTED_SWORD_BLADE_SCHEMATIC)
 			.type(Type.SWORD_BLADE)
+			.compositor(defaultCompositor())
 			.build();
 
 	public static Construct SABER_BLADE = Construct.builder()
 			.addIngredient(IRON)
 			.addIngredient(Schematics.SABER_BLADE_SCHEMATIC)
 			.type(Type.SWORD_BLADE)
+			.compositor(defaultCompositor())
 			.build();
 
 	public static Construct OAK_BINDING = Construct.builder()
 			.addIngredient(OAK)
 			.addIngredient(Schematics.BINDING_SCHEMATIC)
 			.type(Type.BINDING)
+			.compositor(defaultCompositor())
 			.build();
 }
