+++ forgero-core/src/test/java/com/sigmundgranaas/forgero/testutil/Tools.java
@@ -1,11 +1,13 @@
 package com.sigmundgranaas.forgero.testutil;
 
+import static com.sigmundgranaas.forgero.testutil.Name.defaultCompositor;
+
+import java.util.Set;
+
 import com.sigmundgranaas.forgero.core.property.attribute.Category;
 import com.sigmundgranaas.forgero.core.state.Ingredient;
 import com.sigmundgranaas.forgero.core.state.composite.Construct;
 import com.sigmundgranaas.forgero.core.state.upgrade.slot.EmptySlot;
-
-import java.util.Set;
 
 public class Tools {
 	public static Construct IRON_PICKAXE = Construct.builder()
@@ -13,5 +15,6 @@ public class Tools {
 			.addIngredient(Ingredient.of(ToolParts.PICKAXE_HEAD))
 			.addUpgrade(new EmptySlot(0, Types.BINDING, "", Set.of(Category.UTILITY)))
 			.type(Types.PICKAXE)
+			.compositor(defaultCompositor())
 			.build();
 }
