+++ forgero-forge/src/main/resources/forgero.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "com.sigmundgranaas.forgeroforge.mixin",
+  "package": "com.sigmundgranaas.forgero.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
   ],
